reject wif for scanner integrations
